+++ source/windows/debug_layer.h
@@ -15,7 +15,7 @@ void debug_layer_end(void);
 void debug_layer_update_gameplay(void);
 void debug_layer_update_editor(void);
 void debug_layer_close(void);
-void debug_layer_manipulate_entity(transform_t* camera, size_t* selected_entity_slot, int* mouse_over_viewport, level_t* curr_level);
+void debug_layer_manipulate_entity(transform_t* camera, int* selected_entity_slot, int* mouse_over_viewport, level_t* curr_level);
 
 #ifdef __cplusplus
 }
